move and rename useHooks -> useLanguage to new @hooks folder
